hiba_space/requirements.txt
gradio>=4.44.0
llama-cpp-python>=0.2.0
huggingface_hub>=0.20.0
